+++ src-docs/App.vue
@@ -331,6 +331,7 @@ import ifThenElseNullableInitialValue from './json/issues/if-then-else-nullable-
 import ifThenElseNullableInitialValue2 from './json/issues/if-then-else-nullable-initial-value-2.json'
 import multipleRegister from './json/issues/multiple-register.json'
 import childrenDisableState from './json/issues/children-disable-state.json'
+import issue8 from './json/issues/issue-8.json'
 import templates from './json/features/templates.json'
 import editJsonData from './json/features/edit-json-data.json'
 import {isSet} from "../src/helpers/utils.js"
@@ -476,7 +477,8 @@ export default {
           'issue/if-then-else-nullable-initial-value': ifThenElseNullableInitialValue,
           'issue/if-then-else-nullable-initial-value-2': ifThenElseNullableInitialValue2,
           'issue/multiple-register': multipleRegister,
-          'issue/children-disable-state': childrenDisableState
+          'issue/children-disable-state': childrenDisableState,
+          'issue/issue-8': issue8
         }
       },
       example: 'editors/all',