+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.3.22
 
+- Exposed Editor.js
+- Added parent reference to templates
 - Extended editJsonData feature to support arrays with toggle button and dialog functionality
 - Implemented an event listener cleanup system to prevent memory leaks